fisioterapia_api/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
psycopg2-binary==2.9.10
django-cors-headers
django-allauth
dj_rest_auth==7.0.1
requests==2.32.3